+++ src/main/java/Project_3_1_2_v3/Project_3_1_2_v3/service/UserService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface UserService {
     List<User> showUsers();
-    User showUser(Long id);
+    User showUser(int id);
     void save(User user);
-    void update(Long id, User user);
-    void delete(Long id);
+    void update(int id, User user);
+    void delete(int id);
 
 }
